requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# MCP 相关
mcp==0.3.0

# HTTP 客户端
requests==2.31.0
httpx==0.25.2

# 数据处理
pandas==2.1.4
numpy==1.25.2

# 日志和配置
colorlog==6.8.0

# 向量数据库相关（具体依赖可能因实际使用的向量数据库而异）
# 例如：chromadb, weaviate-client, pinecone-client 等

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1

# JSON 处理增强
ujson==5.8.0

# 异步支持
aiofiles==23.2.1
asyncio==3.4.3
